Desktop/RAGPROJECT/pyproject.toml
[project]
name = "ragproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "inngest>=0.5.9",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
